pyproject.toml
[tool.poetry]
name = "crypto-data-ingestion"
version = "0.2.8"
description = "The purpose of this library is facilitate the ingestion of data from the CoinGecko API and store it in cloud/self-hosted object storage."
authors = ["alessandro2702 <alessandro.melo2702@gmail.com>"]
license = "MIT"
classifiers = ["Programming Language :: Python :: 3.11",
               "Topic :: Education",
               "Environment :: Console",
               "Natural Language :: English",
               "Development Status :: 1 - Planning",
               "Operating System :: OS Independent"]
readme = "README.md"

[project.urls]
"Code" = "https://github.com/alessandro2702/crypto-data-ingestion"
"Bug Tracker" = "https://github.com/alessandro2702/crypto-data-ingestion/issues"
"Documentation" = "TBD"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
minio = "^7.2.8"
duckdb = "^1.1.1"
pandas = "^2.2.3"
deltalake = "^0.20.2"
boto3 = "^1.35.36"
fsspec = "^2024.9.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
blue = "^0.9.1"
isort = "^5.13.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.5.34"
mkdocstrings = "^0.26.1"
mkdocstrings-python = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue --check --diff . && isort --check --diff ."
docs = "mkdocs serve"
test = "pytest -s -x --cov=crypto_data_ingestion -vv"
post_test = "coverage html"